move code tests to code dirs
