+++ src/amuse/community/mmams/test_mmams.py
