+++ src/amuse/community/nbody6xx/test_nbody6xx.py
