+++ src/amuse/community/smalln/test_smalln.py
